% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mets}
\alias{mets}
\title{Diagnosis of Metabolic Syndrome}
\usage{
mets(x)
}
\arguments{
\item{x}{the name of the person to say hi to}
}
\value{
the output from code
}
\description{
Diagnosis of Metabolic Syndrome
}
\examples{
mets("Jagadish")
}
