% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Sample data for Analysis}
\usage{
sample_data(path)
}
\arguments{
\item{path}{to filename}
}
\value{
a \code{data.frame} which includes:
\describe{
\item{Gender}{Gender - Male or Female}
\item{Wasist circumference cm}{Waist circumference in cm}
\item{Triglycerides, mg/dL}{Triglycerides in mg/dL}
\item{Fasting plasma glucose, mg/dL}{Fasting plasma glucose in mg/dL}
\item{HDLC, mg/dL}{HDL-cholesterol in mg/dL}
\item{Systolic BP}{Systolic BP in mm of Hg}
\item{Diastolic BP}{Diastolic BP in mm of Hg}
}
}
\description{
A simple dataset to test and run the code
}
\examples{
csv = system.file("extdata","Data.csv", package="metsjuly")
sample_data(csv )
}
