% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Read CSV}
\usage{
sample_read(path)
}
\arguments{
\item{path}{to filename}
}
\value{
a\code{tibble}
}
\description{
Read CSV
}
\examples{
csv = system.file("extdata","Data.csv", package="metsjuly")
sample_read(csv )
}
