% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{An example data frame}
\format{
A data frame with parameters needed to diganose metabolic syndrome.
\describe{
  \item{Gender}{Gender in Male or Female}
  \item{WC}{Wasit Circumference in cm}
  \item{TG}{Triglycerides in mg/dL}
  \item{HDLC}{High Density Lipoprotein Cholesterol in mg/dL}
  \item{SBP}{Systolic BP in mm Hg}
  \item{DBP}{Diastolic BP in mm Hg}
  \item{FPG}{Fasting plasma glucose in mg/dL}
}
}
\usage{
x
}
\description{
A data frame with exact column names as specified.
}
\keyword{datasets}
