% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSyndrome.R
\name{MetabolicSyndrome}
\alias{MetabolicSyndrome}
\title{Diagnosis of Metabolic Syndrome}
\usage{
MetabolicSyndrome(x)
}
\arguments{
\item{x}{x is the data}
}
\description{
This will be used to diagnose Metabolic syndrome
}
