% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSyndrome.R
\name{MetabolicSyndrome}
\alias{MetabolicSyndrome}
\title{Diagnosis of Metabolic Syndrome}
\usage{
MetabolicSyndrome(x)
}
\arguments{
\item{x}{a data frame with column names as exactly specified.}
}
\value{
Yes or No
}
\description{
This package will be used to diagnose metabolic syndrome.The criteria used to diagnose is based on the National Cholesterol Education Program - Adult Treatment Panel -III guidelines (NCEP-ATP-III). It is comprised of multiple parameters which includes waist circumference, systolic blood pressure, diastolic blood pressure, fasting plasma glucose, triglycerides and high-density lipoprotein cholesterol. Hence, this package will be useful in the healthcare settings for easy and quick diagnosis of metabolic syndrome.
}
\examples{
MetabolicSyndrome(x)
}
